Donut_Server/requirements.txt
# Core ML dependencies
torch
transformers
pillow

# FastAPI and server dependencies
fastapi
uvicorn[standard]
python-multipart

# Additional utilities
python-dotenv
aiofiles